img_scrapper_backend/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
idna==2.10
psycopg2==2.8.6
pycodestyle==2.6.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
